Added Gap to Driver in Front/Behind/Leader #21
Took way more code than I first thought, stupid C#, but now it works
