+++ exporter.cs
@@ -12,20 +12,47 @@ namespace F1Manager2024Plugin
         private readonly Dictionary<string, string> _driverFilePaths = new Dictionary<string, string>();
         private readonly Dictionary<string, bool> _headersWritten = new Dictionary<string, bool>();
 
+        public readonly string[] carNames = new string[]
+        {
+                "Ferrari1", "Ferrari2",
+                "McLaren1", "McLaren2",
+                "RedBull1", "RedBull2",
+                "Mercedes1", "Mercedes2",
+                "Alpine1", "Alpine2",
+                "Williams1", "Williams2",
+                "Haas1", "Haas2",
+                "RacingBulls1", "RacingBulls2",
+                "KickSauber1", "KickSauber2",
+                "AstonMartin1", "AstonMartin2",
+                "MyTeam1", "MyTeam2"
+        };
+
+        public int CarsOnGrid = 22;
+
         public void ExportData(string carName, Telemetry telemetry, int i, F1Manager2024PluginSettings Settings)
         {
             if (!Settings.ExporterEnabled || !Settings.TrackedDrivers.Contains(carName)) return; // Return if Exporter isn't Enabled of car isn't Tracked.
             try
             {
-                string trackName = F1ManagerPlotter.GetTrackName(telemetry.Session.trackId);
-                string sessionType = F1ManagerPlotter.GetSessionType(telemetry.Session.sessionType);
+                string trackName = TelemetryHelpers.GetTrackName(telemetry.Session.trackId);
+                string sessionType = TelemetryHelpers.GetSessionType(telemetry.Session.sessionType);
 
                 string basePath = Settings.ExporterPath ?? Path.Combine(
                     Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments),
                     "F1ManagerTelemetry");
 
                 string sessionFolder = Path.Combine(basePath, "exported_data", $"{trackName} {sessionType}");
-                string carFolder = Path.Combine(sessionFolder, String.Join(" ", F1ManagerPlotter.GetDriverFirstName(telemetry.Car[i].Driver.driverId), F1ManagerPlotter.GetDriverLastName(telemetry.Car[i].Driver.driverId)));
+                string carFolder = Path.Combine(sessionFolder, String.Join(" ", TelemetryHelpers.GetDriverFirstName(telemetry.Car[i].Driver.driverId), TelemetryHelpers.GetDriverLastName(telemetry.Car[i].Driver.driverId)));
+
+                // Set the number of cars on the grid.
+                if (telemetry.Car[i].Driver.rpm == 0)
+                {
+                    CarsOnGrid = telemetry.Car.Count(c => c.Driver.rpm > 0);
+                }
+                else
+                {
+                    CarsOnGrid = 22;
+                }
 
                 Directory.CreateDirectory(carFolder);
 
@@ -45,22 +72,27 @@ namespace F1Manager2024Plugin
                 {
                     // Session data
                     ["timestamp"] = DateTime.Now.ToString("yyyy-MM-dd HH:mm:ss.fff"),
-                    ["trackName"] = F1ManagerPlotter.GetTrackName(telemetry.Session.trackId) ?? "",
-                    ["sessionType"] = F1ManagerPlotter.GetSessionType(telemetry.Session.sessionType) ?? "",
+                    ["trackName"] = TelemetryHelpers.GetTrackName(telemetry.Session.trackId) ?? "",
+                    ["sessionType"] = TelemetryHelpers.GetSessionType(telemetry.Session.sessionType) ?? "",
                     ["timeElapsed"] = telemetry.Session.timeElapsed,
 
                     // Driver info
                     ["driverNumber"] = telemetry.Car[i].Driver.driverNumber,
-                    ["driverFirstName"] = F1ManagerPlotter.GetDriverFirstName(telemetry.Car[i].Driver.driverId) ?? "",
-                    ["driverLastName"] = F1ManagerPlotter.GetDriverLastName(telemetry.Car[i].Driver.driverId) ?? "",
-                    ["teamName"] = F1ManagerPlotter.GetTeamName(telemetry.Car[i].Driver.teamId, Settings) ?? "",
-                    ["pitstopStatus"] = F1ManagerPlotter.GetPitStopStatus(telemetry.Car[i].pitStopStatus) ?? "",
+                    ["driverFirstName"] = TelemetryHelpers.GetDriverFirstName(telemetry.Car[i].Driver.driverId) ?? "",
+                    ["driverLastName"] = TelemetryHelpers.GetDriverLastName(telemetry.Car[i].Driver.driverId) ?? "",
+                    ["teamName"] = TelemetryHelpers.GetTeamName(telemetry.Car[i].Driver.teamId, Settings) ?? "",
+                    ["pitstopStatus"] = TelemetryHelpers.GetPitStopStatus(telemetry.Car[i].pitStopStatus) ?? "",
                     ["currentLap"] = telemetry.Car[i].currentLap + 1, // Adjust for index
                     ["turnNumber"] = telemetry.Car[i].Driver.turnNumber,
                     ["position"] = telemetry.Car[i].Driver.position + 1, // Adjust for 0-based index
+                    ["gapToLeader"] = telemetry.Car[i].Driver.GapToLeader,
+                    ["carInFront"] = TelemetryHelpers.GetNameOfCarAhead(telemetry, telemetry.Car[i].Driver.position, i, carNames, CarsOnGrid),
+                    ["gapInFront"] = TelemetryHelpers.GetGapInFront(telemetry, telemetry.Car[i].Driver.position, i, carNames, CarsOnGrid),
+                    ["carBehind"] = TelemetryHelpers.GetNameOfCarBehind(telemetry, telemetry.Car[i].Driver.position, i, carNames, CarsOnGrid),
+                    ["gapBehind"] = TelemetryHelpers.GetGapBehind(telemetry, telemetry.Car[i].Driver.position, i, carNames, CarsOnGrid),
 
                     // Tyres
-                    ["compound"] = F1ManagerPlotter.GetTireCompound(telemetry.Car[i].tireCompound) ?? "",
+                    ["compound"] = TelemetryHelpers.GetTireCompound(telemetry.Car[i].tireCompound) ?? "",
                     ["flTemp"] = telemetry.Car[i].flTemp,
                     ["flDeg"] = telemetry.Car[i].flWear,
                     ["frTemp"] = telemetry.Car[i].frTemp,
@@ -86,10 +118,10 @@ namespace F1Manager2024Plugin
                     ["fuel"] = telemetry.Car[i].fuel,
 
                     // Modes
-                    ["paceMode"] = F1ManagerPlotter.GetPaceMode(telemetry.Car[i].paceMode) ?? "",
-                    ["fuelMode"] = F1ManagerPlotter.GetFuelMode(telemetry.Car[i].fuelMode) ?? "",
-                    ["ersMode"] = F1ManagerPlotter.GetERSMode(telemetry.Car[i].ersMode) ?? "",
-                    ["drsMode"] = F1ManagerPlotter.GetDRSMode(telemetry.Car[i].Driver.drsMode) ?? "",
+                    ["paceMode"] = TelemetryHelpers.GetPaceMode(telemetry.Car[i].paceMode) ?? "",
+                    ["fuelMode"] = TelemetryHelpers.GetFuelMode(telemetry.Car[i].fuelMode) ?? "",
+                    ["ersMode"] = TelemetryHelpers.GetERSMode(telemetry.Car[i].ersMode) ?? "",
+                    ["drsMode"] = TelemetryHelpers.GetDRSMode(telemetry.Car[i].Driver.drsMode) ?? "",
 
                     // Timings
                     ["currentLapTime"] = telemetry.Car[i].Driver.currentLapTime,
@@ -104,7 +136,7 @@ namespace F1Manager2024Plugin
                     ["rubber"] = telemetry.Session.rubber,
                     ["airTemp"] = telemetry.Session.Weather.airTemp,
                     ["trackTemp"] = telemetry.Session.Weather.trackTemp,
-                    ["weather"] = F1ManagerPlotter.GetWeather(telemetry.Session.Weather.weather) ?? ""
+                    ["weather"] = TelemetryHelpers.GetWeather(telemetry.Session.Weather.weather) ?? ""
                 };
 
                 // Write to CSV
